requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
Pillow==9.0.1
sqlparse==0.4.2
